Compile on windows

+++ atire/indexer_param_block_rank.c
@@ -255,7 +255,7 @@ puts("");
 ANT_ranking_function *ANT_indexer_param_block_rank::get_indexing_ranker(long long documents, ANT_compressable_integer *lengths, long *quantization, long long *quantization_bits)
 {
 *quantization = this->quantization;
-*quantization_bits = (long long)(this->quantization_bits == -1 ? 5.4 + 5.4e-4 * sqrt(documents) : this->quantization_bits);
+*quantization_bits = (long long)(this->quantization_bits == -1 ? 5.4 + 5.4e-4 * sqrt((double)documents) : this->quantization_bits);
 
 switch (ranking_function)
 	{
